Move pointer clone and drop functions
