+++ eir-fmm/src/reference_count/drop.rs
@@ -1,4 +1,4 @@
-use super::{super::error::CompileError, pointers, records};
+use super::{super::error::CompileError, pointers, record_utilities};
 use crate::{
     type_information::TYPE_INFORMATION_DROP_FUNCTION_ELEMENT_INDEX,
     types,
@@ -14,14 +14,14 @@ pub fn drop_expression(
 ) -> Result<(), CompileError> {
     match type_ {
         eir::types::Type::ByteString => {
-            drop_pointer(builder, &builder.deconstruct_record(expression.clone(), 0)?)?
+            pointers::drop_pointer(builder, &builder.deconstruct_record(expression.clone(), 0)?)?
         }
         eir::types::Type::Function(_) => todo!(),
         eir::types::Type::Record(record) => {
             builder.call(
                 fmm::build::variable(
-                    records::get_record_drop_function_name(record.name()),
-                    records::compile_record_rc_function_type(record, types),
+                    record_utilities::get_record_drop_function_name(record.name()),
+                    record_utilities::compile_record_rc_function_type(record, types),
                 ),
                 vec![expression.clone()],
             )?;
@@ -54,7 +54,7 @@ pub fn compile_record_drop_function(
     let fmm_record_type = types::compile_record(&eir::types::Record::new(definition.name()), types);
 
     module_builder.define_function(
-        records::get_record_drop_function_name(definition.name()),
+        record_utilities::get_record_drop_function_name(definition.name()),
         vec![fmm::ir::Argument::new("record", fmm_record_type.clone())],
         |builder| -> Result<_, CompileError> {
             for (index, type_) in definition.type_().elements().iter().enumerate() {
@@ -78,35 +78,6 @@ pub fn compile_record_drop_function(
         fmm::types::CallingConvention::Target,
         fmm::ir::Linkage::Weak,
     )?;
-
-    Ok(())
-}
-
-fn drop_pointer(
-    builder: &fmm::build::InstructionBuilder,
-    expression: &fmm::build::TypedExpression,
-) -> Result<(), CompileError> {
-    pointers::if_heap_pointer(builder, expression, |builder| {
-        builder.if_(
-            fmm::build::comparison_operation(
-                fmm::ir::ComparisonOperator::Equal,
-                builder.atomic_operation(
-                    fmm::ir::AtomicOperator::Subtract,
-                    pointers::get_counter_pointer(&builder, expression)?,
-                    fmm::ir::Primitive::PointerInteger(1),
-                )?,
-                fmm::ir::Primitive::PointerInteger(0),
-            )?,
-            |builder| -> Result<_, CompileError> {
-                builder.free_heap(expression.clone())?;
-
-                Ok(builder.branch(fmm::build::VOID_VALUE.clone()))
-            },
-            |builder| Ok(builder.branch(fmm::build::VOID_VALUE.clone())),
-        )?;
-
-        Ok(())
-    })?;
 
     Ok(())
 }
