+++ eir-fmm/src/reference_count/mod.rs
@@ -1,7 +1,7 @@
 mod clone;
 mod drop;
 mod pointers;
-mod records;
+mod record_utilities;
 
 pub use clone::*;
 pub use drop::*;
