+++ eir-fmm/src/reference_count/pointers.rs
@@ -1,5 +1,51 @@
 use super::super::error::CompileError;
 
+pub fn clone_pointer(
+    builder: &fmm::build::InstructionBuilder,
+    expression: &fmm::build::TypedExpression,
+) -> Result<(), CompileError> {
+    if_heap_pointer(builder, expression, |builder| {
+        builder.atomic_operation(
+            fmm::ir::AtomicOperator::Add,
+            get_counter_pointer(&builder, expression)?,
+            fmm::ir::Primitive::PointerInteger(1),
+        )?;
+
+        Ok(())
+    })?;
+
+    Ok(())
+}
+
+pub fn drop_pointer(
+    builder: &fmm::build::InstructionBuilder,
+    expression: &fmm::build::TypedExpression,
+) -> Result<(), CompileError> {
+    if_heap_pointer(builder, expression, |builder| {
+        builder.if_(
+            fmm::build::comparison_operation(
+                fmm::ir::ComparisonOperator::Equal,
+                builder.atomic_operation(
+                    fmm::ir::AtomicOperator::Subtract,
+                    get_counter_pointer(&builder, expression)?,
+                    fmm::ir::Primitive::PointerInteger(1),
+                )?,
+                fmm::ir::Primitive::PointerInteger(0),
+            )?,
+            |builder| -> Result<_, CompileError> {
+                builder.free_heap(expression.clone())?;
+
+                Ok(builder.branch(fmm::build::VOID_VALUE.clone()))
+            },
+            |builder| Ok(builder.branch(fmm::build::VOID_VALUE.clone())),
+        )?;
+
+        Ok(())
+    })?;
+
+    Ok(())
+}
+
 pub fn get_raw_pointer(
     pointer: &fmm::build::TypedExpression,
 ) -> Result<fmm::build::TypedExpression, CompileError> {
@@ -11,22 +57,7 @@ pub fn get_raw_pointer(
     .into())
 }
 
-pub fn is_heap_pointer(
-    pointer: &fmm::build::TypedExpression,
-) -> Result<fmm::build::TypedExpression, CompileError> {
-    Ok(fmm::build::comparison_operation(
-        fmm::ir::ComparisonOperator::NotEqual,
-        fmm::build::bitwise_operation(
-            fmm::ir::BitwiseOperator::And,
-            fmm::build::bit_cast(fmm::types::Primitive::PointerInteger, pointer.clone()),
-            fmm::ir::Primitive::PointerInteger(1),
-        )?,
-        fmm::ir::Primitive::PointerInteger(1),
-    )?
-    .into())
-}
-
-pub fn if_heap_pointer(
+fn if_heap_pointer(
     builder: &fmm::build::InstructionBuilder,
     pointer: &fmm::build::TypedExpression,
     then: impl Fn(&fmm::build::InstructionBuilder) -> Result<(), CompileError>,
@@ -55,7 +86,22 @@ pub fn if_heap_pointer(
     Ok(())
 }
 
-pub fn get_counter_pointer(
+fn is_heap_pointer(
+    pointer: &fmm::build::TypedExpression,
+) -> Result<fmm::build::TypedExpression, CompileError> {
+    Ok(fmm::build::comparison_operation(
+        fmm::ir::ComparisonOperator::NotEqual,
+        fmm::build::bitwise_operation(
+            fmm::ir::BitwiseOperator::And,
+            fmm::build::bit_cast(fmm::types::Primitive::PointerInteger, pointer.clone()),
+            fmm::ir::Primitive::PointerInteger(1),
+        )?,
+        fmm::ir::Primitive::PointerInteger(1),
+    )?
+    .into())
+}
+
+fn get_counter_pointer(
     builder: &fmm::build::InstructionBuilder,
     heap_pointer: &fmm::build::TypedExpression,
 ) -> Result<fmm::build::TypedExpression, fmm::build::BuildError> {
